FIX: Fixed some bugs of db

+++ app/backend/src/database/migrations/3-create-transaction.js
@@ -32,7 +32,7 @@ module.exports = {
         },
       },
       value: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(9,2),
         allowNull: false,
       },
       createdAt: {
